Parse lump invoice line
